Update the JS and add a checkWinner function
Move the winnerLogic inside the gameBoard module.
Create a checkWinner function that declares a winner by looking
at every combo in the winnerLogic and sees if it matches the index
of the gameBoardArray. Also add an else if statement when none
of the combos match (declare a tie).

Create two random players with different names and moves in the
playGame module.

In the addeventlistener, add an if else statement that checks
who has the currentmove and then run checkWinner to see if they have
won.