Fix so reset button is more functional
Remove the player parameter from createPlayer functiona
and instead declare a name variable with a prompt.

Declare two player variables that will act as player
name and move storage. And declare a playerContainer
that will contain these players.
Change the current move to empty.

Add eventlistener on both marker buttons, where
the players will get created for each button.

If the playerContainer is 2 or less, push the player
inside the playerContainer.

In the allBoardSquares addeventlistener, add
another condition to the if statement that says
playerContainer should only have max 2 players.

And update the rest of the code so all connects